requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Google Drive
google-api-python-client==2.139.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Document Processing
pypdf2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3

# Embeddings
sentence-transformers==3.1.1

# Vector Store (FAISS)
faiss-cpu==1.8.0

# LLM - Groq
groq==0.11.0

# Storage & Utils
numpy==1.26.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
tenacity==9.0.0
tiktoken==0.7.0

# Async & Caching
asyncio-redis==0.16.0
diskcache==5.6.3

# Logging
structlog==24.4.0